Add ProductUrl property to Carousel DTOs and entity
Introduced a new `ProductUrl` property to `CarouselDto`,
`CarouselCreateDto`, `CarouselUpdateDto`, `CarouselCreateWithFileDto`,
and `CarouselUpdateWithFileDto` in the `Ecom.Application.DTOs.Carousel`
namespace. The property is a string with a default value of an empty
string and a maximum length of 500 characters.

Updated the `Carousel` entity in the `Ecom.Domain.Entity` namespace to
include the `ProductUrl` property. Modified `CarouselService` to handle
this property during updates.

Added a migration (`20251126154433_intital3`) to include the `ProductUrl`
column in the `Carousels` table. Updated `EcomDbContextModelSnapshot`
and its designer file to reflect the schema changes.